pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "videomass"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "Gianluca Pernigotto", email = "jeanlucperni@gmail.com"}
]
maintainers = [
  {name = "Gianluca Pernigotto", email = "jeanlucperni@gmail.com"}
]
description = "Videomass is a cross-platform GUI for FFmpeg and yt-dlp"
license =  {text = "GNU General Public License v3 (GPLv3)"}
keywords = [
    "videomass",
    "converter",
    "wxpython",
    "yt-dlp",
    "ffmpeg"]
classifiers = [
    "Environment :: X11 Applications :: GTK",
    "Development Status :: 5 - Production/Stable",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: Arabic",
    "Natural Language :: Chinese (Simplified)",
    "Natural Language :: Czech",
    "Natural Language :: Dutch",
    "Natural Language :: English",
    "Natural Language :: French",
    "Natural Language :: German",
    "Natural Language :: Hungarian",
    "Natural Language :: Italian",
    "Natural Language :: Portuguese (Brazilian)",
    "Natural Language :: Russian",
    "Natural Language :: Spanish",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Video :: Conversion",
    "Topic :: Multimedia :: Sound/Audio :: Conversion"
]
dependencies = [
    "wxpython>=4.0.7",
    "pypubsub>=4.0.3",
    "requests>=2.31.0"
]

[project.optional-dependencies]
cli = [
    "yt-dlp",
]
build = [
    "build",
    "hatchling",
    "pip",
    "six",
    "wheel",
    "setuptools",
    "babel",
]
pyinstaller = [
    "pyinstaller",
]

[tool.hatch.build.hooks.custom]  # see hatch_build.py file
# https://github.com/pypa/hatch/discussions/1047
dependencies = [
    "babel"
]

[tool.hatch.version]
path = "videomass/vdms_sys/about_app.py"

[tool.hatch.build.targets.sdist]
include = [
    "videomass",
    "debian",
    "docs/man",
    "develop",
    "tests",
    "AUTHORS",
    "BUGS",
    "CHANGELOG",
    "INSTALL",
    "LICENSE",
    "README.md",  # already included as readme
    "TODO",
    "launcher",
    "requirements",
    "xdg",
    ".gitignore",  # already included by default, needed for auto-excludes
    ".flake8",
    ".pylintrc",
]
exclude = [
    "videomass/data/icons/videomass.icns",
    "videomass/data/icons/videomass.ico",
    "videomass/data/locale/videomass.pot",
    "videomass/data/locale/README",
    "videomass/data/FFMPEG",
]

[tool.hatch.build.targets.wheel]
packages = ["videomass"]
artifacts = [
    # https://hatch.pypa.io/latest/config/build/#artifacts
    "*.mo",
]

[tool.hatch.build.targets.wheel.shared-data]
"docs/man/videomass.1.gz" = "share/man/man1/videomass.1.gz"
"xdg/io.github.jeanslack.videomass.appdata.xml" = "share/metainfo/io.github.jeanslack.videomass.appdata.xml"
"xdg/io.github.jeanslack.videomass.desktop" = "share/applications/io.github.jeanslack.videomass.desktop"
"videomass/data/icons/videomass.png" = "share/pixmaps/videomass.png"
"videomass/data/hicolor/48x48/apps/videomass.xpm" = "share/icons/hicolor/48x48/apps/videomass.xpm"
"videomass/data/hicolor/48x48/apps/videomass.png" = "share/icons/hicolor/48x48/apps/videomass.png"
"videomass/data/hicolor/256x256/apps/videomass.png" = "share/icons/hicolor/256x256/apps/videomass.png"
"videomass/data/hicolor/scalable/apps/videomass.svg" = "share/icons/hicolor/scalable/apps/videomass.svg"

[project.gui-scripts]
videomass = "videomass.gui_app:main"

[project.urls]
Homepage = "https://jeanslack.github.io/Videomass/"
Documentation = "https://jeanslack.github.io/Videomass/Docs.html"
Wiki = "https://github.com/jeanslack/Videomass/wiki"
Repository = "https://github.com/jeanslack/Videomass"
Issues = "https://github.com/jeanslack/Videomass/issues"
